refact: aplica melhorias pedagogicas de constructor injection, pagination, hateoas completo, DTOs e seguranca

+++ src/main/java/com/gs/agroid/dto/LeituraRequestDto.java
@@ -2,7 +2,7 @@ package com.gs.agroid.dto;
 
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
-
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record LeituraRequestDto(
@@ -11,7 +11,7 @@ public record LeituraRequestDto(
 
     @NotNull(message = "O valor da leitura é obrigatório.")
     @Min(value = 0, message = "O valor da leitura não pode ser negativo.")
-    Double valor,
+    BigDecimal valor,
 
     LocalDateTime timestamp
 ) {}
