+++ src/main/java/com/gs/agroid/dto/PropriedadeResponseDto.java
@@ -2,6 +2,7 @@ package com.gs.agroid.dto;
 
 import lombok.*;
 import org.springframework.hateoas.RepresentationModel;
+import java.math.BigDecimal;
 
 @Data
 @EqualsAndHashCode(callSuper = true)
@@ -12,6 +13,6 @@ public class PropriedadeResponseDto extends RepresentationModel<PropriedadeRespo
     private Long id;
     private String nome;
     private String localizacao;
-    private Double tamanho;
+    private BigDecimal tamanho;
     private Long usuarioId;
 }
